requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu129
torch==2.8.0+cu129
torchvision==0.23.0+cu129
einops==0.8.1
kornia==0.8.1
matplotlib==3.9.4
numpy==2.0.2
opencv-python==4.12.0.88
pandas==2.3.1
pillow==11.0.0
transformers==4.55.0
